+++ data/scripts/entity/init.lua
@@ -5,5 +5,6 @@ if onServer() then
         Entity():addScriptOnce("data/scripts/systems/raycast.lua")
         Entity():addScriptOnce("data/scripts/entity/entityAlerts.lua")
         Entity():addScriptOnce("data/scripts/entity/mainCaliber.lua")
+        Entity():addScriptOnce("data/scripts/entity/starfall_setbonuses.lua")
     end
 end